requirements.txt
Django==5.1.6
django-classy-tags==4.1.0 
django-model-utils==4.5.1
django-phonenumber-field==6.1.0
django-polymorphic==3.1.0
django-taggit==6.1.0
django-redirects==0.6.0
phonenumberslite==8.12.48
